refactor(docs-gen): move extension extractor config
